requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
PyPDF2
python-docx
fpdf
wordcloud
